Add Image component
Because the portal manages both a light and dark mode it is often the
case that images need to be swapped out to better suit the colour mode.
To accommodate for this an Image component has been implemented to
handle the logic of swapping out the image reactively when the colour
mode changes.